server_py/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
apscheduler
akshare
tenacity
httpx
simhash
flashtext
sqlite-utils
aiosqlite
sqlmodel
pytest
pytest-asyncio
httpx
pytest-cov
pytest-mock
chromadb
networkx
python-louvain
